% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_smcmc.R
\name{coef.smcmc}
\alias{coef.smcmc}
\title{Extract estimates from \code{smcmc} objects}
\usage{
\method{coef}{smcmc}(object, method = "bayes", nreps = 1000, boot = TRUE,
  quant = 0.9, ...)
}
\arguments{
\item{object}{("smcmc") object.}

\item{method}{if set to "bayes" then the output will be the empirical covariance matrix
of the chain(s). In set to "mle" then a quadratic model will be fitted
through the likelihood evaluations and the output covariance will be the 
inverse of the negative Hessian.}

\item{nreps}{(integer) number of simulations used to tilt the negative Hessian (-H)
toward positive definiteness (PD). Used only if the -H is not PD and \code{method == "asym"}.}

\item{boot}{(logical) relevant only if -H is not PD. 
If TRUE hessians will be simulated by resampling parameters and likelihoods. 
If FALSE hessians will be simulated the asymptotic distribution of the regression
coefficients.}
}
\description{
Extract estimates from \code{smcmc} objects
}
\author{
Matteo Fasiolo <matteo.fasiolo@gmail.com>
}

