% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demvn.R
\name{demvn}
\alias{demvn}
\title{Empirical multivariate Gaussian density}
\usage{
demvn(y, X, log = FALSE, verbose = TRUE)
}
\arguments{
\item{y}{points at which the MVN is evaluated (d dimensional vector) or an n by d matrix, each row indicating
a different position.}

\item{X}{n by d matrix containing the data.}

\item{log}{If TRUE the log of the log-density is returned.}

\item{verbose}{currently not used.}
}
\value{
The density of the empirical MVN, evaluated at y.
}
\description{
Gives a pointwise evaluation of the multivariate normal (MVN) fit to X at position y.
}
\examples{
X <- matrix(rnorm(2 * 1e3), 1e3, 2)
demvn(rnorm(2), X, log = TRUE)
}
\author{
Matteo Fasiolo <matteo.fasiolo@gmail.com> and Simon Wood.
}

