% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsaddle.R
\name{dsaddle}
\alias{dsaddle}
\title{Empirical saddlepoint density}
\usage{
dsaddle(y, X, decay = 0.5, deriv = FALSE, log = FALSE,
  normalize = FALSE, fastInit = FALSE, control = list(),
  multicore = !is.null(cluster), ncores = detectCores() - 1,
  cluster = NULL)
}
\arguments{
\item{y}{points at which the SPA is evaluated (d dimensional vector) or an n by d matrix, each row indicating
a different position.}

\item{X}{n by d matrix containing the data.}

\item{decay}{Rate at which the SPA falls back on a normal density. Should be a positive number,
by default set to 0.5.}

\item{deriv}{If TRUE also the gradient of the log-saddlepoint density is returned.}

\item{log}{If TRUE the log of the saddlepoint density is returned.}

\item{control}{A list of control parameters with entries:
\itemize{
\item{ \code{method} }{The method used to calculate the normalizing constant. 
                       Either "LAP" (laplace) or "IS" (importance sampling).}
\item{ \code{nNorm} }{If control$method == "IS", this is the number of importance samples used.}
\item{ \code{tol} }{The tolerance used to assess the convergence of the solution to the saddlepoint equation.
                    The default is 1e-6.}
}}

\item{tol}{Tolerance used to assess the convergence of the rootfinding routine used to fit
the saddlepoint density. Default value is 1e-6.}
}
\value{
A list with entries:
        \itemize{
        \item{ \code{llk} }{The value of the empirical saddlepoint at y;}
        \item{ \code{mix} }{The mixture of normal-saddlepoint used (1 means only saddlepoint);}
        \item{ \code{grad} }{The gradient of the log-density at y (optional);}
        }
}
\description{
Gives a pointwise evaluation of the empirical saddlepoint and optionally of
             its gradient at position y
}
\author{
Matteo Fasiolo <matteo.fasiolo@gmail.com> and Simon N. Wood.
}
\references{
Fasiolo, M., Wood, S. N., Hartig, F. and Bravington, M. V. (2016). 
            An Extended Empirical Saddlepoint Approximation for Intractable Likelihoods. ArXiv XXX.
}

