% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik_mcmc.R
\name{logLik.mcmc}
\alias{logLik.mcmc}
\title{Extract maximized log-likelihood for \code{mcmc} objects}
\usage{
\method{logLik}{mcmc}(object, burn = 0, quant = 0.1, nreps = 1000,
  boot = TRUE, ...)
}
\arguments{
\item{object}{("mcmc") object.}

\item{nreps}{(integer) number of simulations used to tilt the negative Hessian (-H)
toward positive definiteness (PD). Used only if the -H is not PD and \code{method == "asym"}.}

\item{boot}{(logical) relevant only if -H is not PD. 
If TRUE hessians will be simulated by resampling parameters and likelihoods. 
If FALSE hessians will be simulated the asymptotic distribution of the regression
coefficients.}
}
\description{
Extract maximized log-likelihood for \code{mcmc} objects
}
\author{
Matteo Fasiolo <matteo.fasiolo@gmail.com>
}

