% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstOptim.R
\name{mstOptim}
\alias{mstOptim}
\title{Solving multiple related problems using smart initialization}
\usage{
mstOptim(y, lamHat, objFun, gradFun, mst = NULL, ...)
}
\arguments{
\item{y}{matrix where each row represents the parameters of a different problem}

\item{lamHat}{matrix where each row represents the approximate solution to each problem.
This is used to calculate the distance between problems. For instance the distance
between the i-th and the j-th problem is the Euclidean distance between lamHat[i, ] and lamHat[j, ].
In some cases it is appropriate to choose lamHat = y.}

\item{objFun}{function that solves the problem. I has signature function(.y, .lambda, ...) and it returns a list
that contains the vector "lambda", which is the solution to the problem.}

\item{gradFun}{function that returns a matrix containing the derivatives of lambda wrt y. It has 
signature function(.y, .lambda, ...).}

\item{mst}{(optional) minimum spanning tree which can be provided by the user.}

\item{...}{extra arguments to be passed to objFun() and gradFun()}
}
\description{
Solving multiple related problems using smart initialization
}
\author{
Matteo Fasiolo <matteo.fasiolo@gmail.com>.
}

