% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectDecay.R
\name{selectDecay}
\alias{selectDecay}
\title{Cross-validation for the empirical saddlepoint density}
\usage{
selectDecay(decay, simulator, K, nrep = 1, normalize = FALSE, draw = TRUE,
  multicore = !is.null(cluster), cluster = NULL, ncores = detectCores() -
  1, control = list(), ...)
}
\arguments{
\item{decay}{Numeric vector containing the possible value of the tuning parameter.}

\item{nrep}{Number of times the whole cross-validation procedure will be repeated, by calling
\code{simulator} to generate random variable and computing the cross-validation score
for every element of \code{decay}.}

\item{multicore}{see \code{\link{smcmc-class}}.}

\item{cluster}{an object of class \code{c("SOCKcluster", "cluster")}. This allowes the user to pass her own cluster,
which will be used if \code{multicore == TRUE}. The user has to remember to stop the cluster.}

\item{ncores}{see \code{\link{smcmc-class}}.}

\item{control}{A list of control parameters, with entries:
\itemize{
\item{ \code{K} }{The number of folds to be used in cross-validation. By defaults \code{K = 10};}
\item{ \code{draw} }{ If \code{TRUE} the results of cross-validation will be plotted. \code{TRUE} by default;}
\item{ \code{method} }{The method used to calculate the normalizing constant. 
                       Either "LAP" (laplace) or "IS" (importance sampling).}
\item{ \code{tol} }{The tolerance used to assess the convergence of the solution to the saddlepoint equation.
                    The default is 1e-6.}
\item{ \code{nNorm} }{ Number of simulations to be used in order to estimate the normalizing constant of the saddlepoint density.
                       By default equal to 1e3.}
}}

\item{simulate}{Function with prototype \code{function(...)} that will be called \code{nrep} times
to simulate \code{d}-dimensional random variables. 
Each time \code{simulator} is called it will return a \code{n} by \code{d} matrix.}
}
\value{
A list with entries:
        \itemize{
         \item{ \code{summary} }{ A matrix of summary results from the cross-validation procedure.  }
        \item{ \code{negLogLik} }{A matrix \code{length{decay}} by \code{control$K*nrep} where the i-th row represent the negative loglikelihood
                                  estimated for the i-th value of \code{decay}, while each column represents a different fold and repetition.}
        \item{ \code{normConst} }{ A matrix \code{length{decay}} by \code{nrep} where the i-th row contains the estimates of the normalizing constant.}
        }
        is returned invisibly.
        If \code{control$draw == TRUE} the function will also plot the cross-validation curve.
}
\description{
Performs cross-validation to choose the \code{decay} tuning parameters 
             which determines the mixture between a consistent and a Gaussian estimator
             of the Cumulant Generating Function (ECGF).
}
\examples{

# Saddlepoint is needed
selectDecay(decay = c(0.05, 0.1, 0.5, 1), 
            simulator = function(...) rgamma(100, 2, 1), 
            nrep = 4)
            
# Saddlepoint is not needed
selectDecay(decay = c(0.05, 0.1, 0.5, 1), 
            simulator = function(...) rnorm(100, 0, 1), 
            nrep = 4)

}
\author{
Matteo Fasiolo <matteo.fasiolo@gmail.com>.
}

