% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_definitions.R
\docType{class}
\name{smcmc-class}
\alias{smcmc-class}
\title{\code{smcmc-class}}
\examples{
# Load "synlik" object
data(ricker_sl)

plot(ricker_sl)
 
# MCMC estimation
set.seed(4235)
ricker_sl <- smcmc(ricker_sl, 
                   initPar = c(3.2, -1, 2.6),
                   niter = 50, 
                   burn = 3,
                   priorFun = function(input, ...) 1, 
                   propCov = diag( c(0.1, 0.1, 0.1) )^2, 
                   nsim = 200, 
                   multicore = FALSE)

# Continue with additional 50 iterations
ricker_sl <- continue(ricker_sl, niter = 50)

plot(ricker_sl)       

}
\author{
Matteo Fasiolo <matteo.fasiolo@gmail.com>
}
\references{
Vihola, M. (2011) Robust adaptive Metropolis algorithm with coerced acceptance rate. 
            Statistics and Computing.
}

