% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smcmc.R
\name{smcmc}
\alias{smcmc}
\title{MCMC parameter estimation for objects of class \code{synlik}.}
\usage{
smcmc(object, initPar, niter, nsim, propCov, burn = 0, nchains = 1,
  saddle = FALSE, priorFun = function(param, ...) 0, targetRate = NULL,
  recompute = FALSE, multicore = !is.null(cluster), cluster = NULL,
  ncores = detectCores() - 1, control = list(), ...)
}
\arguments{
\item{object}{An object of class \code{synlik}.}

\item{initPar}{see \code{\link{smcmc-class}}.}

\item{niter}{see \code{\link{smcmc-class}}.}

\item{nsim}{see \code{\link{smcmc-class}}.}

\item{propCov}{see \code{\link{smcmc-class}}.}

\item{burn}{see \code{\link{smcmc-class}}.}

\item{nchains}{number of MCMC chains used, by default 1.}

\item{priorFun}{see \code{\link{smcmc-class}}.}

\item{targetRate}{see \code{\link{smcmc-class}}.}

\item{recompute}{see \code{\link{smcmc-class}}.}

\item{multicore}{see \code{\link{smcmc-class}}.}

\item{cluster}{an object of class \code{c("SOCKcluster", "cluster")}. This allowes the user to pass her own cluster,
which will be used if \code{multicore == TRUE}. The user has to remember to stop the cluster.}

\item{ncores}{see \code{\link{smcmc-class}}.}

\item{control}{see \code{\link{smcmc-class}}.}

\item{...}{additional arguments to be passed to \code{slik} function, see \code{\link{slik}}.}
}
\value{
If \code{nchains == 1} an object of class \code{smcmc} will be returned. Otherwise a list of \code{nchains}
        \code{smcmc} objects will be returned.
}
\description{
MCMC parameter estimation for objects of class \code{synlik}.
}
\details{
When \code{multicore == TRUE} there are two possible behaviours: if \code{nchains == 1} the 
         the parallelism takes place at each step of the MCMC chain (\code{object@simulator} and 
         \code{object@summaries} are called in parallel). If \code{nchains > 1} each chain will be assigned
         to a different node of the cluster, and there is no parallelism within the chains.
}
\author{
Matteo Fasiolo <matteo.fasiolo@gmail.com>, code for adaptive step from the adaptMCMC package.
}
\references{
Vihola, M. (2011) Robust adaptive Metropolis algorithm with coerced acceptance rate. 
            Statistics and Computing.
}

